LayoutTests/fast/text-extraction/debug-text-extraction-form-controls-expected.txt

-- texttree

root
    role=banner
        label='Main Page Title' 'Welcome to Test Page'
    navigation label='Main navigation'
        list
            link 'Section 1'
            link 'Section 2'
    role=main
        section label='Interactive Elements'
            button describedby='This button does nothing' label='Test button' 'Click Me'
            'This button does nothing'
            input uid=… placeholder='Enter text here'
            role=button title='Test action' 'Clickable Div'
        section label='Content with Roles'
            article
                'Article Title'
                'January 1, 2024'
                'This is some article content with highlighted text.'
            label='Related information'
                'Related Links'
                list
                    link 'Example Link'
                    link 'Test Link'
            role=region
                role=status 'Ready'
                button 'Update Status'
            textarea uid=… 'This is a text box'
            contentEditable uid=… role=textbox
                subscript
                    'WebKit home page:'
                    link 'webkit.org'
            form autocomplete=on name=signup
                'Sign up for our newsletter'
                input uid=… placeholder=Username name=username minlength=3 maxlength=20 required
                input uid=… placeholder=Email pattern=[a-z0-9._%+-]+@[a-z0-9.-]+\\.[a-z]{2,}$ name=email required
                input uid=… placeholder='Zip Code' pattern=[0-9]{5} name=zipcode minlength=5 maxlength=5
                input range uid=… name=volume value=42
                input button uid=… name=cancel value=Cancel
                input uid=… name=submit value='Sign Up Now'
            select uid=…
                option value=one 'Number 1'
                option value=two 'Number 2'
                option selected value=three 'Number 3'
            select uid=… multiple
                option selected value=banana
                option selected value=pear
                option value=apple
    'Footer content with'
    role=img label='copyright symbol' '©'
    '2024'

-- html

<body>
    <div role='banner'>
        <h1 aria-label='Main Page Title'>Welcome to Test Page</h1>
    </div>
    <nav aria-label='Main navigation'>
        <ul>
            <li>
                <a>Section 1</a>
            </li>
            <li>
                <a>Section 2</a>
            </li>
        </ul>
    </nav>
    <main role='main'>
        <section aria-label='Interactive Elements'>
            <button aria-describedby='This button does nothing' aria-label='Test button'>Click Me</button>
            This button does nothing
            <input type='text' uid=… placeholder='Enter text here'>
            <div role='button' title='Test action'>Clickable Div</div>
        </section>
        <section aria-label='Content with Roles'>
            <article>
                Article Title
                January 1, 2024
                This is some article content with highlighted text.
            </article>
            <aside aria-label='Related information'>
                Related Links
                <ul>
                    <li>
                        <a>Example Link</a>
                    </li>
                    <li>
                        <a>Test Link</a>
                    </li>
                </ul>
            </aside>
            <div role='region'>
                <div role='status'>Ready</div>
                <button>Update Status</button>
            </div>
            <textarea uid=…>This is a text box</textarea>
            <div uid=… role='textbox' contenteditable>
                <sub>
                    WebKit home page:
                    <a>webkit.org</a>
                </sub>
            </div>
            <form autocomplete='on' name='signup'>
                Sign up for our newsletter
                <input type='text' uid=… placeholder='Username' name='username' minlength=3 maxlength=20 required>
                <input type='text' uid=… placeholder='Email' pattern='[a-z0-9._%+-]+@[a-z0-9.-]+\\.[a-z]{2,}$' name='email' required>
                <input type='text' uid=… placeholder='Zip Code' pattern='[0-9]{5}' name='zipcode' minlength=5 maxlength=5>
                <input type='range' uid=… name='volume' value='42'>
                <input type='button' uid=… name='cancel' value='Cancel'>
                <input type='submit' uid=… name='submit' value='Sign Up Now'>
            </form>
            <select uid=…>
                <option value='one'>Number 1</option>
                <option value='two'>Number 2</option>
                <option value='three' selected>Number 3</option>
            </select>
            <select uid=…>
                <option value='banana' selected>Banana</option>
                <option value='pear' selected>Pear</option>
                <option value='apple'>Apple</option>
            </select>
        </section>
    </main>
    Footer content with
    <span role='img' aria-label='copyright symbol'>©</span>
    2024
</body>

-- json

{
  "type": "root",
  "children": [
    {
      "type": "container",
      "nodeName": "div",
      "role": "banner",
      "children": [
        {
          "type": "container",
          "nodeName": "h1",
          "aria-label": "Main Page Title",
          "children": [
            {
              "type": "text",
              "content": "Welcome to Test Page"
            }
          ]
        }
      ]
    },
    {
      "type": "navigation",
      "nodeName": "nav",
      "aria-label": "Main navigation",
      "children": [
        {
          "type": "list",
          "nodeName": "ul",
          "children": [
            {
              "type": "list-item",
              "nodeName": "li",
              "children": [
                {
                  "type": "link",
                  "nodeName": "a",
                  "children": [
                    {
                      "type": "text",
                      "content": "Section 1"
                    }
                  ]
                }
              ]
            },
            {
              "type": "list-item",
              "nodeName": "li",
              "children": [
                {
                  "type": "link",
                  "nodeName": "a",
                  "children": [
                    {
                      "type": "text",
                      "content": "Section 2"
                    }
                  ]
                }
              ]
            }
          ]
        }
      ]
    },
    {
      "type": "container",
      "nodeName": "main",
      "role": "main",
      "children": [
        {
          "type": "section",
          "nodeName": "section",
          "aria-label": "Interactive Elements",
          "children": [
            {
              "type": "button",
              "nodeName": "button",
              "aria-label": "Test button",
              "aria-describedby": "This button does nothing",
              "children": [
                {
                  "type": "text",
                  "content": "Click Me"
                }
              ]
            },
            {
              "type": "text",
              "content": "This button does nothing"
            },
            {
              "type": "input",
              "nodeName": "input",
              "uid": "…",
              "controlType": "text",
              "placeholder": "Enter text here",
              "children": [
                {
                  "type": "text",
                  "content": "Enter text here"
                }
              ]
            },
            {
              "type": "container",
              "nodeName": "div",
              "role": "button",
              "title": "Test action",
              "children": [
                {
                  "type": "text",
                  "content": "Clickable Div"
                }
              ]
            }
          ]
        },
        {
          "type": "section",
          "nodeName": "section",
          "aria-label": "Content with Roles",
          "children": [
            {
              "type": "article",
              "nodeName": "article",
              "children": [
                {
                  "type": "text",
                  "content": "Article Title"
                },
                {
                  "type": "text",
                  "content": "January 1, 2024"
                },
                {
                  "type": "text",
                  "content": "This is some article content with highlighted text."
                }
              ]
            },
            {
              "type": "container",
              "nodeName": "aside",
              "aria-label": "Related information",
              "children": [
                {
                  "type": "text",
                  "content": "Related Links"
                },
                {
                  "type": "list",
                  "nodeName": "ul",
                  "children": [
                    {
                      "type": "list-item",
                      "nodeName": "li",
                      "children": [
                        {
                          "type": "link",
                          "nodeName": "a",
                          "children": [
                            {
                              "type": "text",
                              "content": "Example Link"
                            }
                          ]
                        }
                      ]
                    },
                    {
                      "type": "list-item",
                      "nodeName": "li",
                      "children": [
                        {
                          "type": "link",
                          "nodeName": "a",
                          "children": [
                            {
                              "type": "text",
                              "content": "Test Link"
                            }
                          ]
                        }
                      ]
                    }
                  ]
                }
              ]
            },
            {
              "type": "container",
              "nodeName": "div",
              "role": "region",
              "children": [
                {
                  "type": "container",
                  "nodeName": "div",
                  "role": "status",
                  "children": [
                    {
                      "type": "text",
                      "content": "Ready"
                    }
                  ]
                },
                {
                  "type": "button",
                  "nodeName": "button",
                  "children": [
                    {
                      "type": "text",
                      "content": "Update Status"
                    }
                  ]
                }
              ]
            },
            {
              "type": "textarea",
              "nodeName": "textarea",
              "uid": "…",
              "controlType": "textarea",
              "children": [
                {
                  "type": "text",
                  "content": "This is a text box"
                }
              ]
            },
            {
              "type": "contentEditable",
              "nodeName": "div",
              "uid": "…",
              "role": "textbox",
              "children": [
                {
                  "type": "subscript",
                  "nodeName": "sub",
                  "children": [
                    {
                      "type": "text",
                      "content": "WebKit home page:"
                    },
                    {
                      "type": "link",
                      "nodeName": "a",
                      "children": [
                        {
                          "type": "text",
                          "content": "webkit.org"
                        }
                      ]
                    }
                  ]
                }
              ]
            },
            {
              "type": "form",
              "nodeName": "form",
              "autocomplete": "on",
              "name": "signup",
              "children": [
                {
                  "type": "text",
                  "content": "Sign up for our newsletter"
                },
                {
                  "type": "input",
                  "nodeName": "input",
                  "uid": "…",
                  "controlType": "text",
                  "placeholder": "Username",
                  "name": "username",
                  "minLength": 3,
                  "maxLength": 20,
                  "required": true,
                  "children": [
                    {
                      "type": "text",
                      "content": "Username"
                    }
                  ]
                },
                {
                  "type": "input",
                  "nodeName": "input",
                  "uid": "…",
                  "controlType": "text",
                  "placeholder": "Email",
                  "pattern": "[a-z0-9._%+-]+@[a-z0-9.-]+\\.[a-z]{2,}$",
                  "name": "email",
                  "required": true,
                  "children": [
                    {
                      "type": "text",
                      "content": "Email"
                    }
                  ]
                },
                {
                  "type": "input",
                  "nodeName": "input",
                  "uid": "…",
                  "controlType": "text",
                  "placeholder": "Zip Code",
                  "pattern": "[0-9]{5}",
                  "name": "zipcode",
                  "minLength": 5,
                  "maxLength": 5,
                  "children": [
                    {
                      "type": "text",
                      "content": "Zip Code"
                    }
                  ]
                },
                {
                  "type": "input",
                  "nodeName": "input",
                  "uid": "…",
                  "controlType": "range",
                  "name": "volume",
                  "value": "42"
                },
                {
                  "type": "input",
                  "nodeName": "input",
                  "uid": "…",
                  "controlType": "button",
                  "name": "cancel",
                  "value": "Cancel"
                },
                {
                  "type": "input",
                  "nodeName": "input",
                  "uid": "…",
                  "controlType": "submit",
                  "name": "submit",
                  "value": "Sign Up Now"
                }
              ]
            },
            {
              "type": "select",
              "nodeName": "select",
              "uid": "…",
              "options": [
                {
                  "value": "one",
                  "label": "Number 1",
                  "selected": false
                },
                {
                  "value": "two",
                  "label": "Number 2",
                  "selected": false
                },
                {
                  "value": "three",
                  "label": "Number 3",
                  "selected": true
                }
              ]
            },
            {
              "type": "select",
              "nodeName": "select",
              "uid": "…",
              "options": [
                {
                  "value": "banana",
                  "label": "Banana",
                  "selected": true
                },
                {
                  "value": "pear",
                  "label": "Pear",
                  "selected": true
                },
                {
                  "value": "apple",
                  "label": "Apple",
                  "selected": false
                }
              ],
              "multiple": true
            }
          ]
        }
      ]
    },
    {
      "type": "text",
      "content": "Footer content with"
    },
    {
      "type": "container",
      "nodeName": "span",
      "role": "img",
      "aria-label": "copyright symbol",
      "children": [
        {
          "type": "text",
          "content": "©"
        }
      ]
    },
    {
      "type": "text",
      "content": "2024"
    }
  ],
  "version": 2
}
